LPS-106546 Move the VerifyResourceActions as an upgrade since we can't reproduce it again after LPS-66047
